requirements.txt
aiohttp==2.3.1
aiohttp-debugtoolbar==0.4.1
aiohttp-devtools==0.6.3
aiohttp-jinja2==0.14.0
aiomysql==0.0.9
aiosmtplib==1.0.2
asn1crypto==0.23.0
async-timeout==2.0.0
cffi==1.11.2
chardet==3.0.4
click==6.7
coverage==4.4.2
cryptography==2.1.3
ecdsa==0.13
gunicorn==19.7.1
idna==2.6
isort==4.2.15
Jinja2==2.10
MarkupSafe==1.0
multidict==3.3.2
peewee==2.10.2
peewee-async==0.5.10
py==1.5.2
pyasn1==0.4.2
pycparser==2.18
pycrypto==2.6.1
PyJWT==1.5.3
PyMySQL==0.7.11
pytest==3.2.3
pytest-aiohttp==0.1.3
pytest-cov==2.5.1
six==1.11.0
watchgod==0.0.3
yarl==0.15.0
